+++ lib/app/modules/survey/views/splash_screen_view.dart
@@ -1,24 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-
-class SurveySplashController extends GetxController {
-  var loading = true.obs;
-  var loadingMessage = "Loading master data...".obs;
-
-  @override
-  void onInit() {
-    // TODO: implement onInit
-    super.onInit();
-    _synchronizeData();
-  }
-
-  void _synchronizeData() async {
-    await Future.delayed(Duration(seconds: 2));
-    loadingMessage.value = "Done!";
-    await Future.delayed(Duration(seconds: 1));
-    loading.value = false;
-  }
-}
+import 'package:sufi_one/app/modules/survey/controllers/survey_splash_controller.dart';
 
 class SurveySplashView extends StatelessWidget {
   final controller = Get.put(SurveySplashController());
